test(root): consolidate 11 scope-isolation-parity binaries into reborn_scope_isolation_suite
Structural move: git mv the 11 tests/reborn_*_scope_isolation_parity.rs
binaries into tests/reborn_scope_isolation_suite/, relocate their shared
mount headers (parity_qa_support, reborn_support, support) to a new
tests/reborn_scope_isolation_suite.rs fat root that #[path]-mounts each
member explicitly, and rewrite each member's internal references to the
shared modules as crate::-qualified. Root-package test targets: 108 -> 98
(30 top-level Rust bins). No test function bodies changed.

Planner: a file inside a root-suite member directory now maps to the
owning fat binary's partition instead of raising "unmapped Reborn test
path". check-test-suite-boundaries.sh's partition guard gained a
suite-member variant (a member referencing a parity_qa_support module
requires its owning tests/<dir>.rs to declare the mount); verified
red/green by temporarily removing the mount and restoring it.
classify-test-scope.sh needed no logic change (its tests/reborn_* glob is
already a prefix match), pinned with a probe.

Equivalence proof: the deduplicated union of all 11 standalone binaries'
`--list` output (30 unique test identities: 11 real parity tests + 19
tests belonging to the shared support modules, which every standalone
binary redundantly relinked and listed) matches the consolidated binary's
`--list` output exactly, modulo the expected module-path qualification
the 11 real tests pick up from now being submodules instead of crate
roots (reborn_agent_scope_isolation_parity -> reborn_agent_scope_isolation_parity::reborn_agent_scope_isolation_parity).
clippy is clean (-D warnings, --all-features). All 30 tests pass (29 ok,
1 pre-existing ignore requiring a live Google credential). Root partition
3 (this suite's owning partition) run end-to-end through
run-hermetic-deterministic-suite.sh's hermetic wrapper under cargo
nextest: 146/146 tests passed.

Two deviations from the plan's literal recipe, both re-verified against
the live worktree:
- Each of the 11 originals carried an `#[allow(dead_code)]` line directly
  above each `#[path]` mount attribute (2 per file); the plan's sed only
  deleted the `#[path]`+`mod` pairs, leaving these orphaned. Extended the
  cleanup to remove them too.
- The plan's equivalence-check sed assumed each standalone binary's
  `--list` output was just its own one test; empirically, all 11 also
  relink and list every test inside the shared support modules they
  mount, so the "before" listing needs deduplication (`sort -u`) before
  comparison, and the "after" listing should NOT have its module-path
  prefix stripped (only the 11 real tests are expected to gain a
  module-qualifier from consolidation; the shared-module tests' existing
  qualifiers must be preserved to compare correctly).

Trace-parity (12 files), QA-phrase (6), and binary-e2e (2) families
remain deferred, gated on Task 6's post-merge compile-share measurement
(recorded in the lane PR, nearai/ironclaw#7799) per the plan's 40%
threshold. Plan JSON schema is unchanged -- the new planner arm only
widens path classification, so in-flight PRs keep working.
