+++ tests/reborn_scope_isolation_suite/reborn_http_network_scope_isolation_parity.rs
@@ -1,11 +1,9 @@
-#[allow(dead_code)]
-#[path = "support/reborn_parity_qa/mod.rs"]
-mod parity_qa_support;
-#[allow(dead_code)]
-#[path = "integration/support/mod.rs"]
-mod reborn_support;
-mod support;
-
+use crate::parity_qa_support::{
+    binary_e2e::RebornBinaryE2EHarness,
+    model_replay::{
+        RebornModelReplayStep, RebornScriptedProviderToolCall, RebornTraceReplayModelGateway,
+    },
+};
 use ironclaw_host_api::{
     action::{NetworkPolicy, NetworkScheme, NetworkTargetPattern},
     ids::CapabilityId,
@@ -13,12 +11,6 @@ use ironclaw_host_api::{
 use ironclaw_host_runtime::HTTP_CAPABILITY_ID;
 use ironclaw_loop_host::{HostManagedModelMessageRole, HostManagedModelResponse};
 use ironclaw_turns::TurnStatus;
-use parity_qa_support::{
-    binary_e2e::RebornBinaryE2EHarness,
-    model_replay::{
-        RebornModelReplayStep, RebornScriptedProviderToolCall, RebornTraceReplayModelGateway,
-    },
-};
 
 #[tokio::test]
 async fn reborn_http_network_scope_isolation_parity() {
