+++ tests/reborn_scope_isolation_suite/reborn_thread_binding_isolation_parity.rs
@@ -1,17 +1,9 @@
-#[allow(dead_code)]
-#[path = "support/reborn_parity_qa/mod.rs"]
-mod parity_qa_support;
-#[allow(dead_code)]
-#[path = "integration/support/mod.rs"]
-mod reborn_support;
-mod support;
-
+use crate::parity_qa_support::binary_e2e::RebornBinaryE2EHarness;
+use crate::parity_qa_support::model_replay::RebornTraceReplayModelGateway;
+use crate::reborn_support::harness::RecordingTestCapabilityPort;
 use ironclaw_loop_host::HostManagedModelResponse;
 use ironclaw_threads::{MessageKind, MessageStatus, ThreadMessageRecord};
 use ironclaw_turns::TurnStatus;
-use parity_qa_support::binary_e2e::RebornBinaryE2EHarness;
-use parity_qa_support::model_replay::RebornTraceReplayModelGateway;
-use reborn_support::harness::RecordingTestCapabilityPort;
 
 #[tokio::test]
 async fn reborn_thread_binding_isolation_parity() {
